Reverting syntaxes as it is confusing users for the time being.
We will need to check each refactoring. As it is integrated.

Not including Type primitives for the time being, and parameters
Removing comments extensions
Removing function parameters

Contract and interface extensions, use the same default colour highlighting
Global, not including generic "." colouring  of functions as it can be confused with another method i.e Transfer.
Not including either global function fields / functions, should be treated the same.
This could be matched if wanted to a normal function.